+++ fixtures/c/thread_local_per_thread.c
@@ -35,7 +35,9 @@ int main() {
     int *handle;
     int *create;
     int *join;
-    int tid;
+    // pthread_t is 8 bytes; under M31 `int` is only 4, so writing
+    // through `&tid` would clobber the next slot. See pthread_create.c.
+    long tid;
     int *retval;
 
     counter = 1;
